exmo/requirements.txt
# This file is part of EXMO2010 software.
# Copyright 2010, 2011 Al Nikolov
# Copyright 2010, 2011 non-profit partnership Institute of Information Freedom Development
# Copyright 2012, 2013 Foundation "Institute for Information Freedom Development"
#
#    This program is free software: you can redistribute it and/or modify
#    it under the terms of the GNU Affero General Public License as
#    published by the Free Software Foundation, either version 3 of the
#    License, or (at your option) any later version.
#
#    This program is distributed in the hope that it will be useful,
#    but WITHOUT ANY WARRANTY; without even the implied warranty of
#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
#    GNU Affero General Public License for more details.
#
#    You should have received a copy of the GNU Affero General Public License
#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
#

Pillow
Django==1.3.7
django-reversion==1.5.7
south
hg+https://bitbucket.org/bsavelev/django-admin-tools
git+https://github.com/django-debug-toolbar/django-debug-toolbar.git
git+https://github.com/playfire/django-debug-toolbar-user-panel.git
git+https://github.com/bsavelev/django403.git
git+https://github.com/bsavelev/django-tagging.git
git+https://github.com/bsavelev/django-tagging-autocomplete.git
django-extensions
feedparser==5.1.2
lxml==2.3.2
django-registration==0.8
django-breadcrumbs==1.1.3-p1
django-ckeditor==3.6.2.1
django-wysiwyg==0.5
simplejson
hg+https://bitbucket.org/offline/django-annoying
python-dateutil
pytils==0.2.3
MySQL-python
